tle2czml/requirements.txt
pygeoif==0.7
python-dateutil==2.6.1
pytz==2018.3
sgp4==1.4
six==1.11.0
wheel==0.24.0
